update risoluzione problemi preferiti

+++ src/main/java/com/project/streaming/service/MoviesService.java
@@ -6,9 +6,12 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+
 import com.project.streaming.model.Movies;
 import com.project.streaming.repository.IMoviesRepository;
 import com.project.streaming.repository.RoleRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class MoviesService {
@@ -31,14 +34,16 @@ public Movies updateMovie(Long id, Movies updatedMovie) {
 
     if (existingMovieOptional.isPresent()) {
         Movies existingMovie = existingMovieOptional.get();
+        existingMovie.setTitle(updatedMovie.getTitle());
         existingMovie.setDescription(updatedMovie.getDescription());
         existingMovie.setAnno(updatedMovie.getAnno());
         existingMovie.setGeneri(updatedMovie.getGeneri());
         existingMovie.setDurata(updatedMovie.getDurata());
         existingMovie.setRating(updatedMovie.getRating());
         existingMovie.setActors(updatedMovie.getActors());
-        existingMovie.setPosterurl(updatedMovie.getPosterurl());
-        existingMovie.setTrailer(updatedMovie.getTrailer());
+        existingMovie.setPoster_url(updatedMovie.getPoster_url());
+        existingMovie.setTrailer_url(updatedMovie.getTrailer_url());
+        existingMovie.setText_png_url(updatedMovie.getText_png_url());
 
         return Moviesdb.save(existingMovie);
     } else {
@@ -46,7 +51,25 @@ public Movies updateMovie(Long id, Movies updatedMovie) {
     }
 }
 
-public void deleteMovie(Long id) {
+//Cancellazione film
+public String deleteMovie(Long id){
+	if(!Moviesdb.existsById(id)) {
+		throw new EntityNotFoundException("film non esiste");
+	}
 	Moviesdb.deleteById(id);
+	return "film Cancellato";
+}
+
+public List<Movies> findByTitle(String title){
+	return (List<Movies>) Moviesdb.findByTitle(title);
+}
+
+//Modifica film
+public Optional<?> putMovies(Movies movie, long id) {
+	if(!Moviesdb.existsById(id)) {
+		throw new EntityNotFoundException("FIlm non esiste");
+	}else {
+		return Optional.of(Moviesdb.save(movie));
+	}
 }
 }
